feat: 리뷰 좋아요 활성화/비활성화 API 구현

+++ src/main/java/com/campfood/src/review/service/ReviewService.java
@@ -3,12 +3,14 @@ package com.campfood.src.review.service;
 import com.campfood.common.error.ErrorCode;
 import com.campfood.common.exception.RestApiException;
 import com.campfood.common.service.EntityLoader;
+import com.campfood.src.member.entity.Member;
 import com.campfood.src.review.dto.request.ReviewCreateDTO;
 import com.campfood.src.review.dto.request.ReviewUpdateDTO;
 import com.campfood.src.review.entity.Review;
+import com.campfood.src.review.entity.ReviewHeart;
 import com.campfood.src.review.entity.ReviewImage;
-import com.campfood.src.review.mapper.ReviewImageMapper;
 import com.campfood.src.review.mapper.ReviewMapper;
+import com.campfood.src.review.repository.ReviewHeartRepository;
 import com.campfood.src.review.repository.ReviewImageRepository;
 import com.campfood.src.review.repository.ReviewRepository;
 import com.campfood.src.store.entity.Store;
@@ -19,6 +21,7 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -27,8 +30,8 @@ public class ReviewService implements EntityLoader<Review, Long> {
 
     private final ReviewRepository reviewRepository;
     private final ReviewImageRepository reviewImageRepository;
+    private final ReviewHeartRepository reviewHeartRepository;
     private final ReviewMapper reviewMapper;
-    private final ReviewImageMapper reviewImageMapper;
 
     private final StoreService storeService;
 
@@ -53,8 +56,15 @@ public class ReviewService implements EntityLoader<Review, Long> {
     @Transactional
     public Long updateReview(final Long reviewId, ReviewUpdateDTO request, List<MultipartFile> reviewImages) {
 
+        // 로그인 유저
+        Member member = null;
+
         // 리뷰 찾기
         Review review = loadEntity(reviewId);
+
+        if (review.getMember().getId().equals(member.getId())) {
+            throw new RestApiException(ErrorCode.UNAUTHORIZED_REVIEW);
+        }
 
         // 리뷰 정보 업데이트
         review.updateReview(request);
@@ -75,8 +85,15 @@ public class ReviewService implements EntityLoader<Review, Long> {
     @Transactional
     public Long deleteReview(final Long reviewId) {
 
+        // 로그인 유저
+        Member member = null;
+
         // 리뷰 찾기
         Review review = loadEntity(reviewId);
+
+        if (review.getMember().getId().equals(member.getId())) {
+            throw new RestApiException(ErrorCode.UNAUTHORIZED_REVIEW);
+        }
 
         // 리뷰 삭제
         review.delete();
@@ -84,10 +101,28 @@ public class ReviewService implements EntityLoader<Review, Long> {
         return review.getId();
     }
 
+    // 리뷰 좋아요 활성화/비활성화 함수
+    @Transactional
+    public boolean toggleReviewHeart(final Long reviewId) {
+
+        // 로그인 유저
+        Member member = null;
+
+        // 리뷰 찾기
+        Review review = loadEntity(reviewId);
+
+        ReviewHeart reviewHeart = reviewHeartRepository.findByMemberAndReview(member, review)
+                .orElse(reviewHeartRepository.save(reviewMapper.toReviewHeart(review, member)));
+
+        reviewHeart.toggleStoreHeart();
+
+        return reviewHeart.isChecked();
+    }
+
     // reviewImage 저장 함수
     private void saveReviewImages(Review review, List<String> imageUrls) {
         List<ReviewImage> reviewImages = imageUrls.stream()
-                .map(imageUrl -> reviewImageMapper.toReviewImage(review, imageUrl))
+                .map(imageUrl -> reviewMapper.toReviewImage(review, imageUrl))
                 .collect(Collectors.toList());
         reviewImageRepository.saveAll(reviewImages);
     }
